requirements.txt
google-adk
google-generativeai
pymongo
gradio